+++ app/Widgets/Presence/Presence.php
@@ -19,6 +19,8 @@ use App\BundleCapabilityResolver;
 use App\Post;
 use App\Widgets\Chats\Chats;
 use App\Widgets\Dialog\Dialog;
+use App\Widgets\Visio\Visio;
+use Movim\CurrentCall;
 use Moxl\Xec\Action\Blocking\Request;
 
 class Presence extends Base
@@ -93,6 +95,10 @@ class Presence extends Base
         $this->rpc('Presence.clearQuick');
 
         $this->user->encryptedPasswords()->delete();
+
+        if (CurrentCall::getInstance()->isStarted()) {
+            //(new Visio)->ajaxEnd(CurrentCall::getInstance()->jid, CurrentCall::getInstance()->id);
+        }
 
         $p = new Unavailable;
         $p->setType('terminate')